fix(wallet): Don't fail in `build_fee_bump` for missing parent txid
This fixes an issue that made using `build_fee_bump` impossible
if the original transaction was created using `add_foreign_utxo`.

Note that it is still required for the previous txouts to exist
in the TxGraph in order to calculate the fee / feerate of the
original transaction, and to populate the witness utxo, etc.

In the future this process could be improved by changing
`add_foreign_utxo` to automatically insert the foreign txout
into the wallet, but to avoid scope creep that change is left
out of this patch.